refactor/ melhoras de codigo
