+++ icasei/src/components/index.ts
@@ -1,2 +1,5 @@
 export * from "./ColorModeSwitcher/ColorModeSwitcher";
 export * from "./Header/Header";
+export * from "./VideoCardForSearch/VideoCardForSearch";
+export * from "./VideoCardForSideBar/VideoCardForSideBar";
+export * from "./VideoPlayer/VideoPlayer";
